Created unit tests for user storage.
To achieve this, user storage has been moved to its own service and decoupled from their direct integration to some controllers. This allows the interface to be tested. Implemented a basic test using an in memory db context (instead of a mock :D) to ensure that retrieval and authentication of users pass checks such as, can the user login, is it deleted, do the credentials decrypt properly? etc.