Null reference fixes
Quick dirty hack, VehicleLocation should be stored in a device handler and not session... session was probably null in this instance which happens when the session is invalidated due to an expired oauth with Nissan. Probably need to rework this further but hopefully this fixes occasional crashes when the app hasn't been used for a while.